+++ CHANGELOG.md
@@ -30,7 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the `config.database.format.month` setting is deprecated in favor of proper three-letter code encoding to support common citation style macros (!34)
 
 ### Fixed
-- renaming the label during the `edit` command does not leave the previous label entry behind
+- renaming the label during the `edit` command does not leave the previous label entry behind:
+    - a followup also ensured that renaming entries happens in-place (#75)
 - the sorting of the `list` command